FIX: move VM
