+++ .env
@@ -1,2 +1,3 @@
 GITHUB_TOKEN=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
 VITE_BASE_URL="http://192.168.0.113:5173"
+VITE_MOCK_API_BASE_URL="http://localhost:8080"
